+++ CHANGELOG.md
@@ -49,6 +49,8 @@ Since version 1.0.0, we try to follow the [Semantic Versioning](https://semver.o
   Set minimum value for _Ambient Light Level_ to 0, if range is not provided. (see [#235](https://github.com/itavero/homebridge-z2m/issues/235))
 - Remove (top level) items with an undefined/null value from state updates.
   This should prevent the warnings mentioned in [#234](https://github.com/itavero/homebridge-z2m/issues/234).
+- `light`: filter properties in state update based on `color_mode`, if provided. (see [#208](https://github.com/itavero/homebridge-z2m/issues/208))
+- `light`: set Hue/Saturation based on Color Temperature (if `color_mode` is also received), to slightly improve the UX. Unfortunately the translation is far from perfect at the moment. (see [#208](https://github.com/itavero/homebridge-z2m/issues/208))
 
 ## [1.3.0] - 2021-06-20
 ### Added